Last harvest code
Potential post-processing feature (for when/if post-processing is added to current pipeline)

Purpose: track when sources were last harvested

The intention of this code is to get the most recent date harvested for each source from our database and update the LastHarvest table to reflect these dates. The current pipeline checks the date of each normalized document to determine whether or not to update the LastHarvest table, which is a functional approach but may result it significant overhead.